+++ tests/Loader/PackageTest.php
@@ -7,15 +7,15 @@ use Innmind\DependencyGraph\{
     Loader\Package,
     Package as Model,
 };
-use Innmind\HttpTransport\Curl;
-use Innmind\TimeContinuum\Earth\Clock;
-use PHPUnit\Framework\TestCase;
+use Innmind\HttpTransport\Transport;
+use Innmind\Time\Clock;
+use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class PackageTest extends TestCase
 {
     public function testInvokation()
     {
-        $load = new Package(Curl::of(new Clock));
+        $load = new Package(Transport::curl(Clock::live()));
 
         $package = $load(Model\Name::of('innmind/url'))->match(
             static fn($package) => $package,
@@ -24,17 +24,17 @@ class PackageTest extends TestCase
 
         $this->assertInstanceOf(Model::class, $package);
         $this->assertSame('innmind/url', $package->name()->toString());
-        $this->assertSame('4.3.2', $package->version()->toString());
+        $this->assertSame('5.1.0', $package->version()->toString());
         $this->assertSame(
             'https://packagist.org/packages/innmind/url',
             $package->packagist()->toString(),
         );
-        $this->assertCount(3, $package->relations());
+        $this->assertSame(1, $package->relations()->size());
     }
 
     public function testMostRecentVersionIsLoaded()
     {
-        $load = new Package(Curl::of(new Clock));
+        $load = new Package(Transport::curl(Clock::live()));
 
         $package = $load(Model\Name::of('guzzlehttp/guzzle'))->match(
             static fn($package) => $package,
@@ -42,12 +42,12 @@ class PackageTest extends TestCase
         );
 
         $this->assertInstanceOf(Model::class, $package);
-        $this->assertSame('7.9.2', $package->version()->toString());
+        $this->assertSame('7.10.0', $package->version()->toString());
     }
 
     public function testReturnNothingWhenThePackageDoesNotExist()
     {
-        $load = new Package(Curl::of(new Clock));
+        $load = new Package(Transport::curl(Clock::live()));
 
         $package = $load(Model\Name::of('psr/http-message-implementation'))->match(
             static fn($package) => $package,