+++ tests/Loader/VendorTest.php
@@ -8,15 +8,17 @@ use Innmind\DependencyGraph\{
     Loader\Package,
     Vendor as Model,
 };
-use Innmind\HttpTransport\Curl;
-use Innmind\TimeContinuum\Earth\Clock;
-use PHPUnit\Framework\TestCase;
+use Innmind\HttpTransport\Transport;
+use Innmind\Time\Clock;
+use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class VendorTest extends TestCase
 {
     public function testInvokation()
     {
-        $http = Curl::of(new Clock)->maxConcurrency(20);
+        $http = Transport::curl(Clock::live())->map(
+            static fn($config) => $config->limitConcurrencyTo(20),
+        );
         $load = new Vendor($http, new Package($http));
 
         $vendor = $load(Model\Name::of('innmind'));
@@ -27,7 +29,7 @@ class VendorTest extends TestCase
             [
                 'innmind/acl',
                 'innmind/amqp',
-                'innmind/ark',
+                'innmind/async',
                 'innmind/async-http-server',
                 'innmind/black-box',
                 'innmind/black-box-symfony',
@@ -35,22 +37,19 @@ class VendorTest extends TestCase
                 'innmind/coding-standard',
                 'innmind/colour',
                 'innmind/crawler',
-                'innmind/crawler-app',
                 'innmind/cron',
                 'innmind/debug',
                 'innmind/dependency-graph',
                 'innmind/di',
-                'innmind/doctrine',
                 'innmind/encoding',
                 'innmind/file-watch',
                 'innmind/filesystem',
+                'innmind/foundation',
                 'innmind/framework',
-                'innmind/genome',
                 'innmind/git',
                 'innmind/git-release',
                 'innmind/graphviz',
                 'innmind/hash',
-                'innmind/homeostasis',
                 'innmind/html',
                 'innmind/http',
                 'innmind/http-authentication',
@@ -59,23 +58,17 @@ class VendorTest extends TestCase
                 'innmind/http-session',
                 'innmind/http-transport',
                 'innmind/immutable',
-                'innmind/infrastructure',
-                'innmind/infrastructure-amqp',
-                'innmind/infrastructure-neo4j',
-                'innmind/infrastructure-nginx',
-                'innmind/installation-monitor',
                 'innmind/io',
                 'innmind/ip',
                 'innmind/ipc',
                 'innmind/json',
                 'innmind/kalmiya',
                 'innmind/lab-station',
-                'innmind/library',
                 'innmind/log-reader',
                 'innmind/logger',
-                'innmind/mantle',
                 'innmind/math',
                 'innmind/media-type',
+                'innmind/mutable',
                 'innmind/object-graph',
                 'innmind/operating-system',
                 'innmind/profiler',
@@ -84,21 +77,13 @@ class VendorTest extends TestCase
                 'innmind/robots-txt',
                 'innmind/router',
                 'innmind/s3',
-                'innmind/scaleway-sdk',
                 'innmind/server-control',
                 'innmind/server-status',
                 'innmind/signals',
-                'innmind/silent-cartographer',
-                'innmind/socket',
                 'innmind/specification',
-                'innmind/ssh-key-provider',
-                'innmind/stack',
                 'innmind/stack-trace',
-                'innmind/stream',
-                'innmind/templating',
-                'innmind/time-continuum',
-                'innmind/time-warp',
-                'innmind/tower',
+                'innmind/static-analysis',
+                'innmind/time',
                 'innmind/type',
                 'innmind/ui',
                 'innmind/url',
@@ -119,12 +104,14 @@ class VendorTest extends TestCase
 
     public function testMaxOpenedFilesRegression()
     {
-        $http = Curl::of(new Clock)->maxConcurrency(20);
+        $http = Transport::curl(Clock::live())->map(
+            static fn($config) => $config->limitConcurrencyTo(20),
+        );
         $load = new Vendor($http, new Package($http));
 
         $vendor = $load(Model\Name::of('symfony'));
 
         $this->assertInstanceOf(Model::class, $vendor);
-        $this->assertCount(250, $vendor->packages());
+        $this->assertSame(350, $vendor->packages()->size());
     }
 }