+++ tests/Package/NameTest.php
@@ -7,7 +7,7 @@ use Innmind\DependencyGraph\{
     Package\Name,
     Exception\DomainException,
 };
-use PHPUnit\Framework\TestCase;
+use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class NameTest extends TestCase
 {